python/src/requirements.txt
Flask == 1.0.3
mysql-connector-python == 8.0.20
prometheus_client == 0.7.1
prometheus_flask_exporter == 0.14.1
configparser
#mysql-python
flask_mysqldb == 0.2.0

#flask-mysqldb
